SEAL-6126 moved logic implementation stuff to a separate project
