+++ modules/flowable-ui-idm/flowable-ui-idm-app/pom.xml
@@ -24,7 +24,10 @@
             <groupId>org.flowable</groupId>
             <artifactId>flowable-ui-idm-logic</artifactId>
         </dependency>
-
+        <dependency>
+            <groupId>org.flowable</groupId>
+            <artifactId>flowable-ui-idm-logic-impl</artifactId>
+        </dependency>
         <dependency>
             <groupId>org.flowable</groupId>
             <artifactId>flowable-idm-engine</artifactId>
