+++ modules/flowable-ui-idm/flowable-ui-idm-app/src/main/webapp/META-INF/context.xml
@@ -16,5 +16,5 @@
 	limitations under the License.
 
 -->
-<Context path="" privileged="true" reloadable="true" crossContext="true">
+<Context crossContext="true" path="/flowable-idm" privileged="true" reloadable="true">
 </Context>
