+++ modules/flowable-ui-idm/flowable-ui-idm-conf/pom.xml
@@ -19,6 +19,10 @@
         </dependency>
         <dependency>
             <groupId>org.flowable</groupId>
+            <artifactId>flowable-ui-idm-logic-impl</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>org.flowable</groupId>
             <artifactId>flowable-ui-idm-rest</artifactId>
         </dependency>
