+++ modules/flowable-ui-idm/flowable-ui-idm-logic-impl/src/main/java/org/flowable/app/idm/service/impl/GroupServiceImpl.java
@@ -10,12 +10,14 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.flowable.app.idm.service;
+package org.flowable.app.idm.service.impl;
 
+import org.flowable.app.idm.service.impl.AbstractIdmService;
 import java.util.List;
 
 import org.apache.commons.lang3.StringUtils;
 import org.flowable.app.idm.constant.GroupTypes;
+import org.flowable.app.idm.service.GroupService;
 import org.flowable.app.service.exception.BadRequestException;
 import org.flowable.app.service.exception.NotFoundException;
 import org.flowable.idm.api.Group;
