+++ modules/flowable-ui-idm/flowable-ui-idm-rest/pom.xml
@@ -17,7 +17,10 @@
             <groupId>org.flowable</groupId>
             <artifactId>flowable-ui-idm-logic</artifactId>
         </dependency>
-
+        <dependency>
+            <groupId>org.flowable</groupId>
+            <artifactId>flowable-ui-idm-logic-impl</artifactId>
+        </dependency>
 		<!-- Flowable -->
 		<dependency>
             <groupId>org.flowable</groupId>
